+++ example/sirius.dart
@@ -3,20 +3,35 @@ import 'package:sirius_backend/sirius_backend.dart';
 void main() {
   Sirius app = Sirius();
 
+  Validator.enableTypeSafety = false;
+
   app.post("/", (Request request) async {
-    // Validator validator = Validator(request.getAllFields(),
-    //     {"name": ValidationRules(defaultValue: "this is defalut value")});
+    Validator validator = Validator(request.getJsonBody, {
+      "email": ValidationRules(required: required(), validEmail: validEmail())
+    });
 
-    // if (!validator.validate()) {
-    //   return Response.send(validator.getError.toString());
-    // }
-    return Response.sendJson({"name": "SOMESH"},
-        headers: {"Content-Type": "text/plain"});
+    if (!validator.validate(typeSafety: true)) {
+      return Response.sendJson(validator.getError.value);
+    }
+    return Response();
   });
 
-  app.start(callback: (server) {
-    print("server is running");
+  app.post("/valid", (Request request) async {
+    Validator validator = Validator(request.getJsonBody, {
+      "email": ValidationRules(required: required(), validEmail: validEmail())
+    });
+
+    if (!validator.validate()) {
+      return Response.sendJson(validator.getError.value);
+    }
+    return Response();
   });
+
+  app.start(
+    callback: (server) {
+      print("server is running");
+    },
+  );
 
   fileWatcher("example/sirius.dart", callback: () async {
     await app.close();