+++ lib/sirius_backend.dart
@@ -4,13 +4,13 @@ library;
 export 'src/sirius.dart';
 export 'src/response.dart';
 export 'src/request.dart';
-export 'src/middleware.dart';
+export 'src/abstract_classes/middleware.dart';
 export 'src/enums/data_types.dart';
 export 'src/enums/date_time_formats.dart';
 export 'src/helpers/file_watcher.dart';
 export 'src/validator.dart';
 export 'src/validation_rules.dart';
-export 'src/wrapper.dart';
+export 'src/abstract_classes/wrapper.dart';
 export 'src/socket_connection.dart';
 
 // TODO: Export any libraries intended for clients of this package.
@@ -20,5 +20,3 @@ export 'src/socket_connection.dart';
 // need to add middleware in web sokets also
 
 // Exception shown in production and error logs creation
-
-// Validator( , , handleException = true/false) for datatype safety